bool varset, alt exp, etc..
expanded the varset ops pattern to include bools

defined flow of ops as 'preloadingt he ops in the varset string, then pushing them into the vars'. IOW outside to inside application

alternat exp symbol, vertical form

planned inversions of math ops to allow reversing th order of non-commutative operations

defined example to show that setting a var can be evaluated and it is equal to the new value of the var.