particle chart work, tweak to scinote
added proton and neutron to the initial particle chart to increase its "universal clarity". Quarks muddy the waters, but protons and neutron ad so much clarity.

fixed particles chart to add tolerances.
uscritpt has a precise definition of what approx-equals means, and the tolerances we have on those quarks don't fit.

lazy hack to invert scinote arrow direction. changed my mind. I want it to 'push the number', not the radix, because then it points in the 'direction of high/low' that matches the effect on the number