subcup types and particle-cats pieces
added 3 new types of subcup brak. these allow explicitification of taxonomy/composition/condition.

the ambiguous sub is still usable, but we need to also be capable of being explicit om this.

particles finally gives the material (objects, cats) to do this

and can use procedural to define conditional, also borrow the subcup component of the if brak as our standalone conditional brak.

to help with clarifying taxonomy, added some extra category symbols for particles

added a tree taxonomy image for definitions

next need to formalize it and actually write out the definitions... tomorrow tomorrow