units decision - not yet implemented though
FINE Hbar G and C=1
I was watching Sabine and I am finally convinced I should 'messy up' the particle mass definition chart.

electron and H21cm are observable units. chart of particles masses and then particles/properties can explain the 21cm line....

but Hbar requires first having units to discuss it(as far as I can tell).

I like introducing units with the unit standard.. It really irks me to introduce a particle mass chart without a 1 anywhere. I loved starting the chart with electron=1 and everything compared to that. and I like introducing the 21cm line and calling its time and length 1.

but...

I also REALLY liked having C=1... which it was and is.(derived from a photon so it was 1)

but it REALLY is awesome for hbar and G to also be 1.... soooo

I will either:

sacrifice the 1's in my unit establishing pages(cringe)

or

let electron mass and 21cm units be the pounds feet and moments of uscript(cringe)
(a moment was 1/40th of an hour)

I do not yet know which idea makes me cringe less